+++ man-roxygen/CohortTableNames.R
@@ -1,4 +1,4 @@
-#' @param cohortDatabaseSchema   Schema name where your cohort table resides. Note that for SQL Server,
+#' @param cohortDatabaseSchema   Schema name where your cohort tables reside. Note that for SQL Server,
 #'                               this should include both the database and schema name, for example
 #'                               'scratch.dbo'.
 #' @param cohortTableNames       List of cohort table names.